@@ -1,6 +1,6 @@
 source 'https://rubygems.org'
 
-gem 'rails', '3.2.12'
+gem 'rails', '4.0.0'
 
 gem 'rails-api'
 
@@ -9,10 +9,9 @@ gem 'airbrake'
 gem 'thin'
 gem 'decent_exposure'
 gem 'active_model_serializers'
-gem 'seedbank'
+gem 'seedbank', github: 'james2m/seedbank'
 gem 'dotenv'
-gem 'strong_parameters'
-gem 'devise'
+gem 'devise', '~> 3.1.0'
 
 group :development do
   gem 'letter_opener'
@@ -43,7 +42,8 @@ group :test do
 end
 
 group :development, :staging do
-  gem 'api_taster', git: 'git@github.com:fredwu/api_taster.git'
+  gem 'api_taster', github: 'vast/api_taster', branch: 'rails4'
+  # gem 'api_taster', path: '../api_taster' # for local development
 end
 
 group :development, :test, :staging do
