@@ -1,5 +1,5 @@
 class ApplicationController < ActionController::API
-  include ActionController::MimeResponds
+  include ActionController::MimeResponds, ActionController::ImplicitRender
 
   self.responder = ApiResponder
 
@@ -8,4 +8,18 @@ class ApplicationController < ActionController::API
   end
 
   respond_to :json
+
+  def authenticate_user!
+    authenticate_user_from_token!
+    super
+  end
+
+  def authenticate_user_from_token!
+    user_token = params[:authentication_token].presence
+    user = user_token && User.where(authentication_token: user_token).first
+
+    if user
+      sign_in user, store: false
+    end
+  end
 end
