@@ -1,9 +1,23 @@
 class User < ActiveRecord::Base
   devise :database_authenticatable, :registerable,
-         :recoverable, :trackable, :validatable,
-         :token_authenticatable
+         :recoverable, :trackable, :validatable
 
-  before_save :ensure_authentication_token
+  before_save :ensure_authentication_token!
 
   has_many :comments
+
+  def ensure_authentication_token!
+    if authentication_token.blank?
+      self.authentication_token = generate_authentication_token
+    end
+  end
+
+  private
+
+  def generate_authentication_token
+    loop do
+      token = Devise.friendly_token
+      break token unless User.where(authentication_token: token).first
+    end
+  end
 end
