@@ -1,1 +1,4 @@
-ApiTaster.global_params = { format: :json } if defined? ApiTaster
+if defined? ApiTaster
+  ApiTaster.global_params = { format: :json }
+  ApiTaster.route_path = Rails.root.join('lib', 'api_tasters')
+end
